requirements.txt
Django==4.2.7
djangorestframework==3.14.0
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
pillow==9.5.0
django-crispy-forms==2.0
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.4.0
dj-database-url==2.0.0
psycopg2-binary==2.9.6
mysqlclient==2.1.1
tensorflow==2.12.0 